# openapi.yaml
openapi: 3.1.0
info:
  title: Planter API
  version: 0.1.0
  description: |
    The Planter API implements the [Phase Manifest Protocol](./PROTOCOL.md).
    It accepts declarative phase plans, diffs them against current state, and executes them.
servers:
  - url: http://localhost:3030
paths:
  /plan:
    post:
      summary: Submit a Phase Manifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Phase'
      responses:
        '200':
          description: Plan received and processed
...
  /phases/{id}:
    get:
      summary: Inspect or rerun a specific phase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Phase details
...
components:
  schemas:
    Phase:
      type: object
      required: [kind, id, spec]
      properties:
        kind:
          type: string
          example: Phase
        id:
          type: string
          example: setup
        spec:
          type: object
          properties:
            description:
              type: string
            selector:
              type: object
              properties:
                matchLabels:
                  type: object
                  additionalProperties:
                    type: string
            waitFor:
              type: object
              properties:
                phases:
                  type: array
                  items:
                    type: string
            retry:
              type: object
              properties:
                maxAttempts:
                  type: integer
            onFailure:
              type: object
              additionalProperties: true